Move CateringOrderTable (refs #243)
It's not used only in reservation/catering form anymore.
