+++ app/shared/catering-order-table/index.js
@@ -1,0 +1,1 @@
+export { default } from './CateringOrderTable';
